requirements.txt
# Stock Market Predictor Dependencies
# For GitHub Actions cloud environment

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Market data
yfinance>=0.2.28

# Machine learning
lightgbm>=4.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0

# Visualization (for performance plots)
matplotlib>=3.7.0

# Web requests (for news sentiment - optional)
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.0